extract makeHot from index.js
