+++ lib/posixify.js
@@ -1,0 +1,3 @@
+module.exports = function posixify(file) {
+	return file.replace(/[/\\]/g, '/');
+};
